requirements.txt
pypdf==5.6.0
PyMuPDF==1.26.1
python-dotenv==1.1.0
langchain-community==0.3.25
langchain_openai==0.3.23
rank_bm25==0.2.2
faiss-cpu==1.11.0
deepeval==3.1.0
llama_index
spacy